Refactor motor and parameter message handling; add light sword and Tesseratica support
- Renamed `respondToParameterMessage` to `handleParameterMessage` in `motors.cpp`, `motors.h`, and `parameterManager.cpp` for consistency.
- Added `confirmParameters` function to `parameterManager.h` to confirm and serialize parameters.
- Implemented accelerometer setup in `sensors.cpp` and included necessary configurations.
- Created `light_sword.cpp` to initialize light sword configurations and sensor states.
- Removed deprecated `slave.cpp` and added `slave.h` to encapsulate slave board functionality.
- Introduced `tess_fragment.cpp` for Tesseratica configurations.
- Cleaned up unused LED and rig structures in `shared.cpp` and `shared.h`.
- Updated `serial.cpp` and `serial.h` to streamline serial management and buffer handling.
- Removed old master implementation in `oldMaster.cpp` as part of codebase cleanup.
