+++ platformio.ini
@@ -36,7 +36,8 @@ lib_deps =
 		moononournation/GFX Library for Arduino@^1.6.0
  
 [env:controller_slave]
-build_flags = -D CONTROLLER_SLAVE  -D CONTROLLER_PINS    -D SLAVE_NAME=\"ControllerSlave\" -D MESH_NET  -D USE_LEDS 
+build_flags = -D CONTROLLER_SLAVE -D SLAVE_BOARD    -D SLAVE_NAME=\"ControllerSlave\"   -D USE_LEDS -D MESH_NET -D CONTROLLER_PINS
+
 lib_deps =
         adafruit/Adafruit MPU6050@^2.2.6
         adafruit/Adafruit NeoPixel@^1.12.2
