+++ src/led_ui/last_device.txt
@@ -1,1 +1,1 @@
-COM1
+COM14
