+++ src/lib/displayManager.h
@@ -8,6 +8,23 @@
 #define DISPLAY_USE_DOUBLE_BUFFER 1
 #endif
 
+#ifdef USE_DISPLAY
+// SSD1306_SWITCHCAPVCC = generate display voltage from 3.3V internally
+if (!display.begin(SSD1306_SWITCHCAPVCC, 0x3C))
+{ // Address 0x3C for 128x32
+    Serial.println(F("SSD1306 allocation failed"));
+    for (;;)
+        ; // Don't proceed, loop forever
+}
+display.display();
+delay(500); // Pause for 2 seconds
+display.setTextSize(1);
+display.setTextColor(WHITE);
+display.setRotation(0);
+
+Serial.println("Adafruit MPU6050 found display!");
+#endif
+
 class DisplayManager
 {
     int _DC_PIN;
