+++ src/lib/stripState.h
@@ -22,7 +22,7 @@ private:
     int stripIndex;
 
     int currentAnimation = 0;
-    int gravityPosition = 0;
+
     int numLEDS = 128;
     int simulateCount = -1;
     int counter = 0;
@@ -71,10 +71,7 @@ public:
     {
         return stripIndex;
     }
-    void setGravityPosition(float position)
-    {
-        gravityPosition = (int)(position * numLEDS);
-    }
+
     void replaceAnimation(int index, ANIMATION_TYPE animType, std::map<ParameterID, float> params = {});
 
     void setLEDRow(LedRow ledRow)
@@ -156,7 +153,7 @@ public:
     {
         return animations;
     }
-    bool respondToParameterMessage(parameter_message parameter);
+    bool handleParameterMessage(parameter_message parameter);
     LED_STATE getLedState() const
     {
         return ledState;
